requirements.txt
streamlit>=1.28.0
googlemaps>=4.10.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
folium>=0.14.0
streamlit-folium>=0.15.0
python-dotenv>=1.0.0
PyDrive2>=2.3.0

